architecture/app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altgraph=0.17.4=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
decorator=4.4.2=pypi_0
expat=2.6.2=h6a678d5_0
freetype=2.12.1=h4a9f257_0
idna=3.7=pypi_0
imageio=2.34.2=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
jpeg=9e=h5eee18b_2
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_1
moviepy=1.0.3=pypi_0
ncurses=6.4=h6a678d5_0
numpy=2.0.1=pypi_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pillow=10.4.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
proglog=0.1.10=pypi_0
pyinstaller=6.9.0=pypi_0
pyinstaller-hooks-contrib=2024.7=pypi_0
python=3.12.3=h996f2a0_0
pytube=15.0.0=pypi_0
pytubefix=6.7.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
setuptools=69.5.1=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.4=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_2
